requirements.txt
python-telegram-bot==20.7
python-dotenv
gspread
oauth2client
nest_asyncio
httpx~=0.25.2
google-auth-oauthlib>=0.4.1
